Restore HPKE
This is close to a complete rework of the HPKE code, because it was
pretty busted when I got here after not being compiled for ages.

Not much needed outside of hpke.rs to make this work, but some.

+++ src/aead/mod.rs
@@ -215,7 +215,7 @@ pub enum AeadAlgorithms {
 
 impl AeadAlgorithms {
     #[must_use]
-    pub const fn key_len(self) -> c_uint {
+    pub const fn key_len(self) -> usize {
         match self {
             Self::Aes128Gcm => 16,
             Self::Aes256Gcm | Self::ChaCha20Poly1305 => 32,
